backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
firebase_admin
python-dotenv
requests
pydantic[email]